services/ingestion-worker/requirements.txt
# Couchbase SDK
couchbase>=4.2.0

# Embeddings
sentence-transformers>=2.2.0
torch>=2.1.0
einops>=0.7.0

# Code parsing
tree-sitter==0.20.4  # Pin to 0.20.x - compatible with tree-sitter-languages
tree-sitter-languages>=1.10.2  # Pre-built language bindings

# GitHub API
PyGithub>=2.1.0
gitpython>=3.1.0

# Document parsing
markdown>=3.5.0
python-frontmatter>=1.0.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0

# Logging
loguru>=0.7.2

# Vector operations
numpy>=1.24.0

# HTTP client for Ollama
httpx>=0.25.0
